exp-track-full-stack/expence_tracker_be/requirement.txt
asgiref==3.8.1
Django==5.2.1
django-cors-headers==4.7.0
django-filter==23.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
psycopg2==2.9.10
PyJWT==2.9.0
sqlparse==0.5.3
typing_extensions==4.14.0
